Fix CIS warning #rb trival #jira none #preflight 61d8e93bc73b7e46b2856b09
#ROBOMERGE-AUTHOR: jimmy.smith
#ROBOMERGE-SOURCE: CL 18553512 in //UE5/Release-5.0/... via CL 18553519
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Staging -> Release-Engine-Test) (v899-18417669)

[CL 18553522 by jimmy smith in ue5-release-engine-test branch]

+++ Engine/Source/Runtime/AudioMixer/Private/AudioMixerSourceManager.cpp
@@ -2942,7 +2942,7 @@ namespace Audio
 
 		// Before adding further commands, ensure we're not growing outside any sensible size for these buffers.
 		// On shipping builds, this will just stop us crashing from growing out of control and OOMing the machine.		
-		const SIZE_T MaxBufferSizeInBytes = CommandBufferMaxSizeInMbCvar << 20;
+		const SIZE_T MaxBufferSizeInBytes = ((SIZE_T)CommandBufferMaxSizeInMbCvar) << 20;
 		if (ensureMsgf(CurrentBufferSizeInBytes < MaxBufferSizeInBytes, TEXT("Command buffer grown to %umb, preventing any more adds! Likely cause, the h/w has stopped consuming data."), CurrentBufferSizeInBytes >>20))
 		{
 			CommandBuffers[AudioThreadCommandIndex].SourceCommandQueue.Add(MoveTemp(InFunction));
